Remove Function from Hooks File
Removed unused console.log
